feat:added components chips
